requirements.txt
certifi==2017.4.17
chardet==3.0.4
defusedxml==0.5.0
Django==1.11.1
django-allauth==0.32.0
django-datetime-widget==0.9.3
django-el-pagination==3.1.0
django-file-form==0.2.3
django-filebrowser==3.7.2
django-filter==1.1.0
django-form-utils==1.0.3
django-grappelli==2.8.3
django-phonenumber-field==1.3.0
django-static-precompiler==1.6
django-tinymce==2.6.0
django-widget-tweaks==1.4.1
easy-thumbnails==2.4.1
idna==2.5
nodeenv==1.1.2
oauthlib==2.0.2
olefile==0.44
pathlib==1.0.1
Pillow==4.1.1
python3-openid==3.1.0
pytz==2017.2
requests==2.18.1
requests-oauthlib==0.8.0
six==1.10.0
urllib3==1.21.1
